requirements.txt
pydantic
pydantic-settings
azure-storage-queue
azure-identity
